add all

+++ database/migrations/2021_10_30_163910_create_ingredient_potion_table.php
@@ -17,9 +17,8 @@ class CreateIngredientPotionTable extends Migration
             $table->id();
             $table->foreignId('ingredient_id')->constrained()->onDelete('cascade');
             $table->foreignId('potion_id')->constrained()->onDelete('cascade');
-            $table->integer('amount');
+            $table->float('amount', 8, 2);
             $table->timestamps();
-
         });
     }
